Documentation format fixes
