+++ DataAccess/DataAccess.Postgre/PostgreRepository.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
+#include <DataAccess/IRepository.hpp>
 #include <Network/Primitives.hpp>
 #include <iostream>
 
-#include "DataAccess/IRepository.hpp"
 #include "PostgreTable.hpp"
 #include "Utility/Exception.hpp"
 #include "Utility/Utility.hpp"
@@ -23,12 +23,12 @@ public:
 
     virtual ~PostgreRepository() = default;
 
-    std::vector<Network::ChannelInfo> getAllChannelsList() override final;
-    std::vector<Network::MessageInfo> getMessageHistoryForUser(const std::uint64_t channelID) override final;
-    std::uint64_t                     loginUser(const std::string& login, const std::string& pwdHash) override final;
-    Utility::StoringMessageCodes      storeMessage(const Network::MessageInfo& mi) override final;
-    Utility::RegistrationCodes        registerUser(const Network::RegistrationInfo& ri) const override final;
-    Utility::DeletingMessageCodes     deleteMessage(const Network::MessageInfo& mi) override final;
+    std::vector<Network::ChannelInfo> getAllChannelsList() final;
+    std::vector<Network::MessageInfo> getMessageHistoryForUser(const std::uint64_t channelID) final;
+    std::uint64_t                     loginUser(const std::string& login, const std::string& pwdHash) final;
+    Utility::StoringMessageCodes      storeMessage(const Network::MessageInfo& mi) final;
+    Utility::RegistrationCodes        registerUser(const Network::RegistrationInfo& ri) const final;
+    Utility::DeletingMessageCodes     deleteMessage(const Network::MessageInfo& mi) final;
 
 private:
     std::optional<pqxx::result> insertMessageIntoMessagesTable(const Network::MessageInfo& msi);
