+++ DataAccess/DataAccess.SQLite/CMakeLists.txt
@@ -11,7 +11,4 @@ set(SOURCE_FILES
 add_library(${TARGET} STATIC ${SOURCE_FILES})
 
 target_link_libraries(${TARGET}
-    PUBLIC Qt5::Sql)
-
-target_link_libraries(${TARGET}
-    PUBLIC DataAccess.Public Utility.Public)
+    PUBLIC Qt5::Sql DataAccess.Public Utility.Public)
