+++ DataAccess/Public/CMakeLists.txt
@@ -2,11 +2,14 @@ set(TARGET DataAccess.Public)
 set(OUTPUT_NAME Juniorgram.DataAccess.Public)
 
 set(SOURCE_FILES
-    Include/DataAccess/IAdapter.hpp
-    Include/DataAccess/IRepository.hpp)
+        Include/DataAccess/IAdapter.hpp
+        Include/DataAccess/IRepository.hpp
+        Include/DataAccess/QueryBuilder.hpp
+        Include/DataAccess/SQLStatements.hpp)
 
 add_library(${TARGET} INTERFACE)
 
 target_sources(${TARGET} INTERFACE ${SOURCE_FILES})
+
 target_include_directories(${TARGET}
-    INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
+        INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
